% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_site.R
\name{launch_site}
\alias{launch_site}
\title{Launch the local version of the website}
\usage{
launch_site()
}
\description{
Just run the command \code{launch_site()} to open a local version of the
course website
}
